move : PartyDocumentException document 패키지로 이동
