+++ src/main/java/com/playus/twpservice/domain/party/exception/document/PartyDocumentException.java
@@ -1,4 +1,4 @@
-package com.playus.twpservice.domain.party.exception;
+package com.playus.twpservice.domain.party.exception.document;
 
 import java.io.Serial;
